[AIRFLOW-XXX] Remove verbose test on nosetests (#4913)
Something about the tests or how we run them changed and we ended up 
with a lot more lines appearing in the output, taking us over Travis' "will 
display in the UI" limit, making it harder to debug failures. This isn't a long
term fix, but improves things while we fix the tests for the better.